Phase 2 Full Stack Todo/backend/requirements.txt
fastapi[standard]>=0.115.0
passlib[bcrypt]>=1.7.4
psycopg2-binary>=2.9.11
pyjwt>=2.9.0
python-multipart>=0.0.21
sqlmodel>=0.0.31
python-dotenv>=1.0.0
uvicorn>=0.30.0
cron-validator>=1.0.8
